fix matchup running in cron-runner instead of bancho (+ dupe bancho client instances)
