+++ BanchoBot/functions/tournaments/matchup/runMatchup.ts
@@ -1,8 +1,9 @@
 import { randomUUID } from "crypto";
+import { banchoClient } from "../../..";
 import { leniencyTime, ScoringMethod } from "../../../../Models/tournaments/stage";
 import { Matchup } from "../../../../Models/tournaments/matchup";
 import { StageType } from "../../../../Models/tournaments/stage";
-import { banchoClient, osuClient } from "../../../../Server/osu";
+import { osuClient } from "../../../../Server/osu";
 import { BanchoChannel, BanchoLobby, BanchoLobbyPlayer, BanchoLobbyTeamModes, BanchoLobbyWinConditions } from "bancho.js";
 import { convertDateToDDDHH, osuLogTimestamp } from "../../../../Server/utils/dateParse";
 import { MappoolMap } from "../../../../Models/tournaments/mappools/mappoolMap";
