+++ BanchoBot/index.ts
@@ -1,10 +1,23 @@
+import { BanchoClient } from "bancho.js";
+import { config } from "node-config-ts";
 import { handleCommand } from "./commands";
-import { banchoClient } from "../Server/osu";
 
-const self = banchoClient.getSelf();
+import baseServer from "../Server/baseServer";
+import koaBody from "koa-body";
+import Mount from "koa-mount";
+
+import banchoRouter from "../Server/api/routes/bancho";
+
+import ormConfig from "../ormconfig";
+
+// Bancho Client
+const banchoClient = new BanchoClient({ username: config.osu.bancho.username, password: config.osu.bancho.ircPassword, botAccount: config.osu.bancho.botAccount, apiKey: config.osu.v1.apiKey });
+banchoClient.connect().catch(err => {
+    if (err) throw err;
+});
 
 banchoClient.on("connected", () => {
-    console.log(`osu! Bancho Bot connected as ${self.ircUsername}`);
+    console.log(`Logged into osu! as ${banchoClient.getSelf().ircUsername}`);
 });
 
 banchoClient.on("PM", async (message) => {
@@ -23,3 +36,21 @@ banchoClient.on("PM", async (message) => {
         await handleCommand(commandName, message, ...args);
     }
 });
+
+const koa = baseServer;
+
+koa.use(koaBody());
+
+// General
+
+/// Cron
+koa.use(Mount("/api/bancho", banchoRouter.routes()));
+
+ormConfig.initialize()
+    .then(async (connection) => {
+        console.log(`Connected to the ${connection.options.database} database!`);
+        koa.listen(config.banchoBot.port);
+    })
+    .catch((error) => console.log("An error has occurred in connecting.", error));
+
+export { banchoClient };
