+++ Dockerfile
@@ -27,6 +27,7 @@ RUN npm run build:mca
 RUN npm run build:main
 RUN npm run build:api
 RUN npm run build:discord-bot
+RUN npm run build:bancho-bot
 
 ENV NODE_ENV=production
 RUN npm prune --production
